Cherry-pick all recent changes.
--------------------------------------
Add timestamp metadata to Entities.

This CL adds metadata to Entity objects in entity.dart. This metadata
includes the timestamps currently implemented in the Suggestinator's
entity_timestamper, which will be removed in a future CL.

R=thatguy@google.com

--------------------------------------

Improve SemanticNode creation performance.

This fixes the problem detailed in issue #822. Local changes to the
StateGraph are now stored via Graph.applyMutations() and not
Graph.mutate(), avoiding unnecessary creation of intermediate
LazyClonedGraph objects that get leaked.

R=thatguy@google.com

--------------------------------------

Add how to sync Dart dependencies to HACKING.md.

R=ianloic@google.com

--------------------------------------

Updates the indexer to support the updating of an existing JSON index.

These changes are part of the effort to provide an indexing service for
the CDN. In particular, we provide 'update' and 'remove' operations on
the Index such that we can work with only the existing index.json file
and the provided manifest instead of re-indexing from scratch.

R=thatguy@google.com, alhaad@google.com
BUG=

--------------------------------------

Instead of keeping a (stale) copy of mojo's mojo/services/ and pulling a different subset of the mojo repo (mojo/public) as mojo_sdk, just do a sparse checkout of the mojo repo with those two directories.

R=alhaad@google.com

--------------------------------------

Fix syncing to reorged repo

TBR=alhaad

Roll flutter, mojo_devtools, mojo_sdk

flutter = 7a9fe01cb99da467da5577f95377149be8e74778
mojo_devtools = 91948c54058447914ff2f3a62f9ae4d6051f0886
mojo_sdk = 6b5fb1227c742f5ecc077486ebc029f2711c61fa
mojo = 969b831b9a6916cd34e41cbaa51e9655d82d9ca2

R=alhaad@google.com

--------------------------------------

Implements deploy script and hello-world for notification handler.

This CL primarily implements the deployment script for the GAE project.
A consequence of using a Dart project with local dependencies is that
a number of dependencies cannot be retrieved using `pub get`.
Furthermore, Docker requires that all files are within the project
directory when the project is to be deployed. The deploy script
achieves this by reorganizing the project folder to correspond to that
suggested in the Docker image, google/dart-runtime-base, and deploys it
to the GAE project using gcloud.

R=alhaad@google.com, thatguy@google.com
BUG=

--------------------------------------

Fix Inspector Session graph.

This CL adds a couple of checks in dataflow.js to prevent the Inspector
from crashing when a given pathExpr is not in inputMatches.

BUG=
R=armansito@google.com

--------------------------------------

Adding basic breadth-first data adaptation

This change makes the Suggestinator bundle headless data adaptation
steps into compound suggestions with an interactive module that would
otherwise have unsatisfied dependencies. This allows module authors to
write data adaptation modules to improve interop without degrading user
experience.

Further work is needed to compose friendlier suggestion descriptions and
to cover additional edge cases, though the latter may be better deferred
until consensus on a data flow design.

BUG=
R=mesch@google.com

--------------------------------------

Implements the notification handling service for the cloud indexer.

This CL implements the notification-handler service, which uses the
indexer pipeline to update the appropriate index given a new manifest.
Note that this implementation makes the assumption that URIs in the
manifest are prefixed with 'https://tq.mojoapps.io', with the path of
the URL corresponding to the path in cloud storage relative to the
appropriately versioned directory.

In a follow-up CL, we will refactor global variables in both default
and notification-handler services to be part of deployment
configuration.

BUG=
R=alhaad@google.com, qsr@google.com, thatguy@google.com

--------------------------------------

Steps towards viewing event log in the Inspector.

This CL sets up a new InspectorJSONServer for the event log. The raw event JSON data can be viewed at localhost:1843/suggestinator/events. Eventually, this data will be viewable in a more useful format using the actual Inspector front-end.

Also includes some general refactoring work I did along the way.

BUG=
R=armansito@google.com

--------------------------------------

Simple Inspector UI for viewing Modular events.

This CL adds front-end components for viewing the event log. Currently, events are only generated in the Suggestinator and the Suggestinator has its own JSON server for publishing those events to the Inspector. To accommodate having multiple JSON servers (the Handler also has one), there is a dropdown menu for switching servers on the right side of the main toolbar. Finally, the cards displaying information about events are relatively lacking right now (e.g. Suggestions are shown in terms of steps added and removed, where each step is just a URL), but can be extended to include whatever information we may deem useful to see.

R=armansito@google.com

-----------------------
Remove line that was accidentally committed.

Updates the Cloud Indexer deploy script to use the Command pattern.

Additionally, this CL refactors both default and notification-handler
services so that the bucket name can be passed through an environment
variable.

BUG=
R=alhaad@google.com, thatguy@google.com

Review URL: https://chromereviews.googleplex.com/485297016 .

Change-Id: Ia2ad2fc43dea20e40ac0f030610a949c745217f9