Hookup device-runner->story-manager->story-runner.
To run this workflow:
$ application_manager mojo:device-runner-boot user1 mojo:example-recipe

This patch adds:
(1) A new device-runner-boot binary which uses [user] (from command line) to
launch a new story-manager. It then calls the |StartNewStory| on the
story-manager with [recipe] (from command line).
(2) Story manager implements 2 interfaces; one meant to be called from
device shell to get the user id and another meant to be called from
sysui to act on individual stories.
(3) Updated gn/apps.

MultiPart: 1/2
Change-Id: I31cd3a5c9d65834cb40c21b045e0bd5ec8dd8e32