Namespace simplifications:
Using a namespace per repository for modular.
This reduces stutter, as we have fewer classes with the same
name as their namespace. Also, we need fewer qualified names
or using declarations.

Add using declarations for all uses of names from the mojo::
namespace in the modular namespace. This makes code less verbose,
and mojo names should be familiar to the reader of the code.

Rename directory application to mojo, and add a README.md to
explain the purpose.

Collateral cleanup:
 * clang-format.
 * drop useless expliciti qualifiers.

Change-Id: If4fd762fc46b2580bd57314d1578fb6360becd14