Store session data in ledger again.
Ledger has fixed two issues that prevented us from reading back
story data stored in the story page:
* https://fuchsia.atlassian.net/browse/LE-63
* https://fuchsia.atlassian.net/browse/LE-46

Add back ability to StoryStorageImpl to read from and write to
a ledger page. Retain ability to store story data in ledger, in
case we might want to be able to switch off ledger in the future.

With this change, the device_runner log demonstrates that a story
can be stopped from the dummy user shell such that when it's started
again it resumes at the state it was stopped at.

Collateral cleanup:

* TransactionContainer is tied to the instance it's a member of,
  so it can just be a member directly.

Change-Id: I6928511287d66acebbdc90151474514c2a193c51