Factor data model from examples. Fix duplication problems on restart.
The Store now broadcasts change events.
The concept of "dirty with respect to the view" is separated from
"dirty with respect to the link."

The dupe problems are fixed by adjusting the application logic in
the modules (not sending a counter with a value less than the current)
and recipe (not sending the first notification).

Remaining TODO items:
- Move Counter out of store.cc.
- Fix the spinning cubes in graphics mode.
- Factor the spinning cube into its own class.
- Figure out how to trigger the animation based on Link values.

Change-Id: Ib787899b69f9ed87898847275a51fc8f5628a718