Adjust applications and their view hosts to the changed view initialization protocol as implemented by flutter:
* example_recipe, example_module1, example_module2 and story_impl,
* dummy_user_shell and user_runner,
* dummy_device_shell and device_runner.

Main challenge is that the request to initialize the view,
CreateView(), and the request to create and then initialize
the primary service instance, are no longer in guaranteed sequence.

As a result, the primary service can no longer directly
implement the view. Instead, the view is implemented as a
separate class, which is instanted for the CreateView request.

As a simplification, we assume that there is only a single
primary service instance per application instance (which is
true for all modules, since we always start a new module
instance with CreateApplication()), so the App instance
can implement the primary service instance, as well as
the service and view providers.

Collateral cleanup:

* Drop using declarations for fidl and modular.
* Move dummy implementations out of the modular namespace.

Change-Id: Ia1ff1cf600cbe35d00cd7e235b4f9d145a2da107