Support '-' argument for bootstrap, to specify no app.
This is useful for debugging in the presence of the @<scope>
environment selector mechanism in the shell, since we can launch
additional apps within the scope of the bootstrap environment
after the fact.  We still require bootstrap to take an argument
since not launching anything within the bootstrapped environment
would generally be a mistake in a production system.

Change-Id: Ib66460d8a853138751a380e1b87ebc5cdbf84011