DevUserShell
DevUserShell is an implementation of UserShell that runs a story with
a single module. It takes two flags:

 --root-module - The URL of the module to run.
 --root-link - value for the Link instance of the module, as JSON.

Run it like this:

  @ bootstrap device_runner --user-shell=dev_user_shell --user-shell-args=--root-link={"foo":"bar"\,"baz":1}

Notice the comma in the JSON escaped by \. Also notice there must be no
space in the JSON value anywhere.

Fix the earlier submitted args parser to actually include characters
that are escaped in the output.

Add a very basic converter from JSON to Link Values. To be improved.

FW-63

Change-Id: I71778b483b2755b78c234e00e225eb10fd606f69