Provide a per-module ledger instance via Story interface.
Story.GetLedger() returns ledger for the particular module under the user.
The same module run under a different story will also access the same ledger
(consequently, multiple instances of this module running will access the
same ledger). It is important to use transactions for this reason.

To accomplish this, this CL does the following:
- Have UserRunner manage LedgerRepositoryFactory.
- Encapsulate the LedgerRepositoryFactory into an object and pass plumb
  a reference to it to StoryProviderImpl and StoryController (both of
  which live under the UserRunner app).
- Have StoryRunner.CreateStory take in a LedgerRepository for the user.
  StoryController will vend a new LedgerRepository to each StoryRunner.
- Add a GetLedger() to Story. StoryImpl+StoryConnection vend
  module-specific ledgers.

To exercise this code:
- example_recipe.cc (run by dummy_user_shell) manages a counter in its
  root ledger page, that gets printed+incremented everytime the module
  recipe is run.

Change-Id: I6955de8bc0af0ef727821a2d9934504f4a545b66