Resolve the race condition in DeleteStory().
When a DeleteStory() request is received, block new requests
for the story controller and queue further delete requests.

In the watcher notification for a deleted story, ensure that
the story controller is first stopped before the story
controller is deleted.

Concurrent requests to resume and delete a story are still
possible, but deletion takes precedence.

Also return null from a GetStoryInfo() request for a non-
existing story, rather than to just not return anything.
Close the StoryController request connection in ResumeStory()
for a non-existing story.

Collateral cleanup and fixes:

* Also fix that StoryController.Stop() does nothing if no
  story is running. (Previously it would crash.)

* ResumeStory() doesn't attempt to create a StoryController
  if the story ID doesn't exist, but closes the request
  outright.

* Brush up comments.

Change-Id: Ib1398bc11e7fcb2fdf440a25ed2269e98a8ac8a7