Fix the StoryController instance life cycle once more.
There are two ways in which a StoryController instance can be
torn down:

1. The story is deleted by DeleteStory() locally or by notification
   about remote deletion from the ledger.

2. By disconnecting all clients.

Each time, PurgeControllers() is called, and the StoryController
instance is deleted if one of the two conditions is met, but
nothing towards meeting the other condition is in progress. If it is,
purging is deferred until the operation in progress finishes (at
which time PurgeControllers() will be called again).

Change-Id: I234d40288e3c2c4dff23504e2ddb69464b1e9925