One more fix in the StoryController life cycle.
If a story controller stops its story because it was deleted,
then it must not write any story data to the ledger, because
that would bring back the record for the deleted story, undeleting
it.

Also, while a story controller still exists while it tears down
its story when being deleted, it can in principle still get calls
from its client to start again and to get info. We also prevent
these invocations, but we need a more systematic way to handle
state, TBD.

Change-Id: I3d5a941a4147e951d8dd0d179dfbc74d928f7fa9