Use an environment supplied ApplicationLoader.
With this change ApplicationEnvironments can supply their own
ApplicationLoader by providing a service. I experimented initially
adding an ApplicationLoader argument to CreateNestedEnvironment and then
adding a GetApplicationLoader method to ApplicationEnvironmentHost, but
in both cases it became necessary to add a Duplicate method to
ApplicationLoader and/or add more GetApplicationLoader methods all over
the place.

Then I remembered that this is the whole kind of reason we have
ServiceProviders so went that route and everything is a lot cleaner.

Also, this depends on: https://fuchsia-review.googlesource.com/c/17349/
to allow applications to be launched from a VMO rather than a path. The
ApplicationLoader interface doesn't know or care about paths.

Change-Id: Ib53ace93c17687ab7280cc3524174f8d340533de