Make it more natural to run non-Module applications in a story.
If an application doesn't implement the Module service, it can still
be run in a Story. We indicate that by the new Module state UNLINKED,
because such an application would not access other modules in the
story through Links.

We still want to distinguish that state from the situation that the
application cannot be run in the first place (for example, because
it doesn't exist), which we still indicate with the ERROR state.

This does not translate to a new Story state: At the story level,
a story with an UNLINKED root module is just RUNNING.

We use the new state to improve shutdown of a module: If we know
the module won't respond to a Stop() call, we can stop it outright,
and don't need to time out on waiting for it.

Change-Id: I0120bdaef1b673a3dfd80bcd52ebc20c3eaab6d8