ComponentContext: Provide ComponentContext interfaces via Story.GetComponentContext()
This interface is vended (and implementation is scoped) from StoryConnection,
so that the life time of this ComponentContext is per-module-instance.

Also in this CL:
- Move ComponentContext.Done() and put it in AgentContext; this will
allow Done() impls to be custom to agent and module components.
- Move agent_controller.fidl into a sub-directory, since fidl dart
targets are packed per-directory, so we can't have two fidl targets in
the same GN build file.

Change-Id: Ifc1911b84a7dacb065ba1528c251b5f6cc193662