If a story is read back from the ledger, it's by definiton not running.
This is a quick hack to get user shell to be able to start stories
that were running when it got killed and the is_running flag is kept
on in the ledger. The real fix involves to actually keep story state
per device, but we don't have this yet.

FW-95

Change-Id: I9b3f531641fdfd8355ad42459fa8f0f0c699099e