Add a run sequence diagram for modular.
Starting from kernel, this diagram shows the sequence of binaries that
get run in response to "@ bootstrap device_runner".
It also shows two running user runners and stories to demonstrate the
ApplicationEnvironment where they run in.
NOTE: This is how things currenly look.

Change-Id: I0d6ef1c2d39c73c4ce814750ce8f14c3299df13d
Nextup: A new canvas for - which application provides what services to whom.