TestRunner: Change semantics of the Store, update child modules to use it
This changes the semantic of the Store to be a producer-consumer queue (each key is a
different queue), as suggested by mesch@ in a previous discussion. This
semantic is much better when trying to model "waiting for signals".

Other changes in this CL:
- I updated the tests to use this model, and removed the timed module
stops -- things should happen causally now. I did leave a 5 second
timeout for the test to fail.
- I made the "App" classes heap-allocated and self-owned, so that they
can properly fail on their TestPoint when the test is torn down: right
now, there is a race between being-torn-down-by-test-runner and
exiting-test-process.

Once we write a few more of these tests, we should come up with some
clean macros and helper classes/functions to distill these patterns.

Change-Id: Ic91f107a9b892511f1192d937caba614caeb308a