Implement tasks which run in response to triggers.
This patch implements MessageQueue based triggers; i.e. a new task will
get run on an agent (via Agent.RunTask()) if a message is received on a
MessageQueue that was registerd using AgentContext.ScheduleTask().
If the agent was not running, triggering a task will cause an agent to
be started in a new process and Agent.Initialize() be called before
Agent.RunTask() is called.

A task continues to run until the callback from Agent.RunTask() => ()
is completed. In the future we want to cap this behavior via strategies like
budgets.

An agent is stopped (via Agent.Stop()) when there are no active
connections and all tasks have completed.

This is tested by creating a module, which connects to an agent and receievs
a MessageQueue token from the agent. The module then stops the agent by
dropping its AgentController. Once the agent has stopped, the module
sends it a message; which trigers a task on it.

Change-Id: Ibaf53549a381511366dcefc81467699a36172aa0