Drop() Operation instance before result callback.
Instead of calling the result callback and then Done(), the result
callback is passed to the Operation() constructor, and Done() is
called with the argument to be passed to the result callback. The
result callback is invoked only after the Operation is deleted. The
Operation now takes the type of result callback argument as template
parameter.

This doesn't work by just reversing the order of Done() and the result
callback, because Done() invokes Run() on the next Operation,
so the next operation would be invoked before the result callback of the
previous operation. That reverses order of operations in the queue.

Therefore, we split Done() in DoneStart(), which is called before the
result callback and deletes the Operation instance, and DoneFinish(),
which is called after the result callback and calls Run() on the next
Operation if there is one, though the newly introduced method Cont()
of OperationContainer.

It may happen that the result callback causes the operation container
to get deleted. To detect that, DoneStart() returns a weak pointer
to the operation container, and passes it back to DoneFinish(). DoneFinish()
only calls Cont() if the operation container still exists.

This does anything only for operation queues, because an operation
collection always runs its operation immediately, so there never is a
next operation to run after one finishes.

Also, the result callback might enqueue more operations, but they must
not be run before the result callback finishes. This is detected by
checking the weak pointer factory for outstanding wek pointers.

Change-Id: I57ae9bcc503f4c7f2da9d188b11440b50476ad76