Fix running of tests/test.sh.
This was broken because the run_modular_test script waits for a
particular string in the log output, and only when it receives it does
it go to run the next test. Fix by doing 2 things-
(1) Print 'test_runner: teardown <test_id>' in the log once test
completes.
(2) Run test_runner in bootstrap directly instead of in virtual console.
This is because virtual console does not forward log to log listener.

Change-Id: Ib58094aee81547a589f74e493f2dcbf9c0f3aac5