Preliminary auth flow.
* Plumb auth token from device shell -> device runner -> user runner ->
env service.
* Add a flag to device_runner which can get auth tokens from a config
file.

Notes:
* It currently requires more work to get webview into the default
fucshia build. There are many reason for this (ask me offline).
* Until we have webview for auth, and even afterwards for ease of
development, there is flag in device runner 'device_runner
--use_auth_config_file' that reads auth tokens from a file which is put
into the filesystem. Instructions for generating this config file are at
https://fuchsia.googlesource.com/modules/+/master/#Email. Currently this
config file contains token for a single user, but in the future we want
to extend this to multiple users.

Change-Id: Ibb1d81886c3047a849ec225098f1ecba1a4587da