Pass view_id, parent_id, and view_type to StoryShell.ConnectView()
Adds a module instance id for each started module, so that the story can
associate an id with each view. ModuleContext keeps track of the id for the
module it is connected to so it can be passed into story_impl when needed. With
module ids for the calling module and the started module, we can pass a view_id,
view_id of the parent module, and view_type for the StoryShell.

Also updates StartRootModule() to use the more generic StartStoryInShell()
method.

Change-Id: I4fd7a6ceaaadd6f6c790af98ed0c76f1fcfb66bb